% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_series_models.R
\name{FitModels}
\alias{FitModels}
\title{Fit time series models for the data of the credit indicators.}
\usage{
FitModels(data, horizon)
}
\arguments{
\item{data}{Should be the data of the credit indicators returned by the function \code{\link{ReadData}}.}

\item{horizon}{Integer that represents the number of months that will be considered to generate the forecast based on the model results.}
}
\value{
A list of objects containing the model results, correlation between the observed and fitted values, horizon, Indicator and Index. Also, a dataset containing the fitted values and the forecast results (lower and upper limits of the 95\% confidence interval) is provided.
}
\description{
This function fits the best ARIMA model and provides a forecast for data of the credit indicators.
}
\examples{
data0 <- ReadData(Indicator = 'Interest rate', Index = 'Total')
model <- FitModels(data=data0, horizon=20)

}
\seealso{
\code{\link{ReadData}}, \code{\link{auto.arima}}, \code{\link{forecast}} and \code{\link{cor}}.
}
