% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_series.R
\name{plot_FitModels}
\alias{plot_FitModels}
\title{Visualisation of the credit indicator}
\usage{
plot_FitModels(data)
}
\arguments{
\item{data}{Should be the data of the credit indicators returned by the function \code{\link{FitModels}}.}
}
\description{
This function provides visualisations of the credit indicator (observed data, fitted values and lower and upper limits of the 95\% confidence interval for the forecast).
}
\examples{
data0 <- ReadData(Indicator = 'Interest rate', Index = 'Total')
model <- FitModels(data=data0, horizon=20)
plot_FitModels(model)
}
\seealso{
\code{\link{ReadData}} and \code{\link{FitModels}}.
}
