% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{ReadData}
\alias{ReadData}
\title{Databases of the credit indicators from the Central Bank of Brazil.}
\usage{
ReadData(Indicator, Index)
}
\arguments{
\item{Indicator}{The credit indicator can be either: Allowance, Due loans, Due loans by capital, Spread, House debt and Interest rate.}

\item{Index}{Given the Indicator, the parameter Index identifies which information should be consider. The parameter Index varies depending on the Indicator (see below Details).}
}
\value{
A list of objects containing the dataset with the time series associated to the Indicator and Index, a vector with the dates of the times series, the Indicator and the Index.
}
\description{
This function reads and manipulates the databases of the credit indicators of the public and private banks from Brazil. The information was obtained in the site of the Central Bank of Brazil.
}
\details{
For instance, when Indicator = 'Allowance', the parameter Index can be Total, Public banks, National private control, Foreign control and Private control.
\itemize{
 \item{Total: }{Percentage of total provision in relation to the loan portfolio of the financial system.}
 \item{Public banks: }{Percentage of total provision in relation to the loan portfolio of the financial institutions under public control.}
 \item{National private control: }{Percentage of total provision in relation to the loan portfolio of the financial institutions under national private control.}
 \item{Foreign control: }{Percentage of total provision in relation to the loan portfolio of the financial institutions under foreign control.}
 \item{Private control: }{Percentage of total provision in relation to the loan portfolio of the financial institutions under private control.}
}

For Indicator = 'Spread', the parameter Index can be Total, New operations companies, New operations households, Non-revolving total, Non-revolving companies and Non-revolving households.
\itemize{
 \item{Total: }{Average spread of new credit operations - Total - (percentage points) p.p.}
 \item{New operations companies: }{Average spread of new credit operations - Companies - Total - p.p.}
 \item{New operations households: }{Average spread of new credit operations - Households - Total - p.p.}
 \item{Non-revolving total: }{Average spread of new non-revolving credit operations - Total - p.p.}
 \item{Non-revolving companies: }{Average spread of new non-revolving credit operations - Companies - Total - p.p.}
 \item{Non-revolving households: }{Average spread of new non-revolving credit operations - Households - Total - p.p.}
}

For Indicator = 'House debt', the parameter Index can be Total and Without mortgage.
\itemize{
 \item{Total: }{Household debt - \%.}
 \item{Without mortgage: }{Household debt without mortgage loans - \%.}
}

For Indicator = 'Interest rate', the parameter Index can be Total, New operations companies, New operations households, Non-revolving total, Non-revolving companies and Non-revolving households.

\itemize{
 \item{Total: }{Average interest rate of new credit operations - Total - \% (per year) p.y.}
 \item{New operations companies: }{Average interest rate of new credit operations - Companies - Total - \% p.y.}
 \item{New operations households: }{Average interest rate of new credit operations - Households - Total - \% p.y.}
 \item{Non-revolving total: }{Average interest rate of new non-revolving credit operations - Total - \% p.y.}
 \item{Non-revolving companies: }{Average interest rate of new non-revolving credit operations - Companies - Total - \% p.y.}
 \item{Non-revolving households: }{Average interest rate of new non-revolving credit operations - Households - Total - \% p.y.}
}
}
\examples{
data0 <- ReadData(Indicator = 'Interest rate', Index = 'Total')
data1 <- ReadData(Indicator = 'Spread', Index = 'New operations companies')
data2 <- ReadData(Indicator = 'Household debt', Index = 'Without mortgage')
data3 <- ReadData(Indicator = 'Allowance', Index = 'Public banks')
}
\seealso{
\code{\link{FitModels}} and \code{\link{plot_FitModels}}.
}
